Finished input field error checking and contact page validation
